Put all the logic inside the actor
