+++ src/cast_core/src/DbCheckActor.cpp
@@ -32,44 +32,139 @@
 #include <gennylib/Cast.hpp>
 #include <gennylib/MongoException.hpp>
 #include <gennylib/context.hpp>
-#include <gennylib/dbcheck.hpp>
+#include <gennylib/v1/Topology.hpp>
 
 
 namespace genny::actor {
 
+namespace {
+bool clearHealthLog(v1::Topology& topology) {
+    class ClearHealthLogVisitor : public v1::TopologyVisitor {
+    public:
+        void onReplSetMongod(const v1::MongodDescription& desc) override {
+            mongocxx::client client(mongocxx::uri(desc.mongodUri));
+            auto local = client["local"];
+            auto healthLog = local["system.healthlog"];
+            healthLog.drop();
+            BOOST_LOG_TRIVIAL(debug) << "Finished clearing healthLog on node: " << desc.mongodUri;
+            _success = true;
+        };
+
+        bool success() {
+            return _success;
+        }
+
+    private:
+        bool _success = false;
+    };
+
+    ClearHealthLogVisitor clearHealthLogVisitor;
+    topology.accept(clearHealthLogVisitor);
+    return clearHealthLogVisitor.success();
+}
+
+bool waitForDbCheckToFinish(v1::Topology& topology) {
+    class WaitForDbCheckVisitor : public v1::TopologyVisitor {
+    public:
+        void onReplSetMongod(const v1::MongodDescription& desc) override {
+            BOOST_LOG_TRIVIAL(debug) << "Waiting for dbcheck to finish on node: " << desc.mongodUri;
+
+            mongocxx::client client(mongocxx::uri(desc.mongodUri));
+            auto local = client["local"];
+            auto healthLog = local["system.healthlog"];
+            // Dbcheck end health log.
+            auto query = bsoncxx::builder::stream::document{} << "operation"
+                                                              << "dbCheckStop"
+                                                              << bsoncxx::builder::stream::finalize;
+            int count = 0;
+            while (!count) {
+                count = healthLog.count_documents(query.view());
+            }
+
+            BOOST_LOG_TRIVIAL(debug) << "dbcheck finished on node: " << desc.mongodUri;
+            _success = true;
+        };
+
+        bool success() {
+            return _success;
+        }
+
+    private:
+        bool _success = false;
+    };
+
+    WaitForDbCheckVisitor waitForDbCheckVisitor;
+    topology.accept(waitForDbCheckVisitor);
+    return waitForDbCheckVisitor.success();
+}
+
+bool dbcheck(mongocxx::pool::entry& client,
+             mongocxx::database& db,
+             std::string& collName,
+             const bsoncxx::document::value& dbCheckCmd) {
+    /* Only one thread needs to actually do the dbcheck. If
+     * another thread enters the function while a dbcheck
+     * is happening, it waits until the dbcheck finishes
+     * and exits.
+     */
+    static std::mutex dbcheckLock;
+    static std::atomic_bool success;
+
+    if (dbcheckLock.try_lock()) {
+        // We are the thread actually running dbcheck.
+        const std::lock_guard<std::mutex> lock(dbcheckLock, std::adopt_lock);
+        v1::Topology topology(*client);
+        bool clearHealthLogSuccess = clearHealthLog(topology);
+        // Run dbcheck.
+        db.run_command(dbCheckCmd.view());
+        bool waitForDbCheckToFinishSuccess = waitForDbCheckToFinish(topology);
+        success = clearHealthLogSuccess && waitForDbCheckToFinishSuccess;
+    } else {
+        // We are waiting for another thread to do the dbcheck.
+        dbcheckLock.lock();
+        dbcheckLock.unlock();
+    }
+
+    return success;
+}
+
+}  // anonymous namespace
+
 struct DbCheckActor::PhaseConfig {
     PhaseConfig(PhaseContext& phaseContext, mongocxx::database&& db, ActorId id)
         : database{db}, 
-          collectionName{phaseContext["Collection"].to<std::string>()},
-          validateMode{phaseContext["ValidateMode"].maybe<std::string>()}  {
+          collectionName{phaseContext["Collection"].to<std::string>()} {
         auto threads = phaseContext.actor()["Threads"].to<int>();
         if (threads > 1) {
             std::stringstream ss;
             ss << "DbCheckActor only allows 1 thread, but found " << threads << " threads.";
             BOOST_THROW_EXCEPTION(InvalidConfigurationException(ss.str()));
         }
+
+        bsoncxx::builder::stream::document dbCheckCmdStream;
+        dbCheckCmdStream << "dbCheck" << collectionName;
+        // Adding parameters.
+        auto validateMode = phaseContext["ValidateMode"].maybe<std::string>();
+        if (validateMode) {
+            dbCheckCmdStream << "validateMode" << validateMode.value();
+        }
+
+        dbCheckCmd = (dbCheckCmdStream << bsoncxx::builder::stream::finalize);
+        BOOST_LOG_TRIVIAL(debug) 
+            << "About to run dbcheck command: " << bsoncxx::to_json(dbCheckCmd.view());
     }
 
     mongocxx::database database;
     std::string collectionName;
-    std::optional<std::string> validateMode;
+    bsoncxx::document::value dbCheckCmd = bsoncxx::from_json("{}");
 };
 
 void DbCheckActor::run() {
     for (auto&& config : _loop) {
         for (const auto&& _ : config) {
             auto dbcheckCtx = _dbcheckMetric.start();
-            auto param = config->validateMode? 
-                bsoncxx::builder::stream::document() << "validateMode"
-                << config->validateMode.value() << bsoncxx::builder::stream::finalize
-                : bsoncxx::builder::stream::document() << bsoncxx::builder::stream::finalize;
-
-            BOOST_LOG_TRIVIAL(debug)
-                << " DbCheckActor with paramters " << bsoncxx::to_json(param.view());
-
             BOOST_LOG_TRIVIAL(debug) << "DbCheckActor starting dbcheck.";
-
-            if (dbcheck(_client, config->database, config->collectionName, param)) {
+            if (dbcheck(_client, config->database, config->collectionName, config->dbCheckCmd)) {
                 dbcheckCtx.success();
             } else {
                 dbcheckCtx.failure();
