--- src/gennylib/include/gennylib/dbcheck.hpp
@@ -1,30 +1,0 @@
-// Copyright 2021-present MongoDB Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include <mongocxx/pool.hpp>
-#include <string>
-
-#include <gennylib/context.hpp>
-
-namespace genny {
-
-/**
- * Helper function to run dbcheck on a replica set and waiting for it to finish.
- */
-bool dbcheck(mongocxx::pool::entry& client,
-             mongocxx::database& db,
-             std::string& collName,
-             bsoncxx::document::value& dbCheckParam);
-
-}  // namespace genny
